+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.8.2",
   "main": "dist/index.js",
   "module": "dist/index.js",
-  "types": "dist/types/src/index.d.ts",
+  "types": "dist/types/index.d.ts",
   "files": [
     "dist",
     "src"
