+++ src/loaders/zarr/ome-zarr.ts
@@ -3,6 +3,39 @@ import { loadMultiscales } from './lib/utils';
 import ZarrPixelSource from './pixel-source';
 import { getLabels } from '../utils';
 
+interface Channel {
+  active: boolean;
+  color: string;
+  label: string;
+  window: {
+    min?: number;
+    max?: number;
+    start: number;
+    end: number;
+  };
+};
+  
+interface Omero {
+  channels: Channel[];
+  rdefs: {
+    defaultT?: number;
+    defaultZ?: number;
+    model: string;
+  };
+  name?: string;
+};
+  
+interface Multiscale {
+  datasets: { path: string }[];
+  version?: string;
+};
+  
+export interface RootAttrs {
+  omero: Omero;
+  multiscales: Multiscale[];
+};
+
+
 export async function load(store: ZarrArray['store']) {
   const { data, rootAttrs } = await loadMultiscales(store);
   const labels = getLabels('XYZCT');
